# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NairaMaster.ng Micro Task Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all active tasks
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task (admin or user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /tasks/{taskId}/renew:
    post:
      operationId: renewTask
      tags: [tasks]
      summary: Renew a task so all users can perform it again (admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}/extend:
    post:
      operationId: extendTask
      tags: [tasks]
      summary: Extend how many performers a task allows (admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskExtendInput"
      responses:
        "200":
          description: Task extended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}/perform:
    post:
      operationId: performTask
      tags: [tasks]
      summary: Mark a task as being performed by the user
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task marked as performed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # --- SUBMISSIONS ---
...
  /updates:
    get:
      operationId: getUpdates
      tags: [updates]
      summary: Get all updates
      responses:
        "200":
          description: Updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Update"
    post:
      operationId: createUpdate
      tags: [updates]
      summary: Create an update (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Update"
  /updates/{updateId}/comments:
    post:
      operationId: addUpdateComment
      tags: [updates]
      summary: Add a comment to an update
      parameters:
        - name: updateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # --- BANNERS ---
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, description, link, socialMedia, category, tags, maxPerformers, performerCount, status, uploadedByAdmin, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        link:
          type: string
        socialMedia:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        maxPerformers:
          type: integer
        performerCount:
          type: integer
        status:
          type: string
          enum: [active, expired, deleted]
        uploadedByAdmin:
          type: boolean
        uploadedByUserId:
          type: ["integer", "null"]
        reward:
          type: ["number", "null"]
        performedByCurrentUser:
          type: boolean
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, description, link, socialMedia, category, tags]
      properties:
        title:
          type: string
        description:
          type: string
        link:
          type: string
        socialMedia:
          type: string
          enum: [facebook, whatsapp, tiktok, instagram, linkedin, pinterest, x, google, youtube, others]
        category:
          type: string
          enum: [like, follow, comment, share, visit, signup, login, survey, optional, others]
        tags:
          type: array
          items:
            type: string
        maxPerformers:
          type: integer
        reward:
          type: number
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        link:
          type: string
        socialMedia:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        maxPerformers:
          type: integer
        reward:
          type: number
        status:
          type: string
    TaskExtendInput:
      type: object
      required: [additionalPerformers]
      properties:
        additionalPerformers:
          type: integer

    # --- Submissions ---
    Submission:
      type: object
      required: [id, taskId, userId, status, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, declined]
        screenshots:
          type: array
          items:
            type: string
        declineReason:
          type: ["string", "null"]
        task:
          $ref: "#/components/schemas/Task"
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Update:
      type: object
      required: [id, title, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
    UpdateInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
    Comment:
      type: object
      required: [id, content, userId, userName, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        userId:
          type: integer
        userName:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # --- Banners ---
...